Don't create asm.frg file if it already exists, could overwrite users data.

+++ src/AS_RUN/runCA-OBT/runCAutil/preoverlap.pl
@@ -59,9 +59,11 @@ sub preoverlap {
         
     $numFrags = getNumberOfFragsInStore($bin, $wrk, $asm);
 
-    if (runCommand($wrk, "$bin/gatekeeper -dumpfrg $wrk/$asm.gkpStore 2> $wrk/gatekeeper.err | grep -v 'No source' > $wrk/$asm.frg")) {
-        unlink "$wrk/$asm.frg";
-    }    
+    if ( ! -s "$wrk/$asm.frg" ) { # don't overwrite if it's already there
+        if (runCommand($wrk, "$bin/gatekeeper -dumpfrg $wrk/$asm.gkpStore 2> $wrk/gatekeeper.err | grep -v 'No source' > $wrk/$asm.frg")) {
+            unlink "$wrk/$asm.frg";
+        }    
+    }
 
  stopafter:
     stopAfter("initialStoreBuilding");
